This Directory consists of files which illustrates and demonstates concepts in;
What Is A “C” Library? What Is It Good For? Creating A Static “C” Library Using “ar” and “ranlib” Using A “C” Library In A Program What is difference between Dynamic and Static library(Static and Dynamic linking)

What is a static library, how does it work, how to create one, and how to use it Basic usage of ar, ranlib, nm